+++ src/components/index.ts
@@ -1,22 +1,14 @@
-export { Input } from './input';
-export { Button } from './button';
 export { Icons } from './icon';
 export { PrivateRoute } from './private-route';
 export { Sidebar, SidebarLink } from './sidebar';
 export { Toolbar } from './toolbar';
-export { Toggler } from './toggler';
 export { PageHeader } from './page-header';
 export { Table, Column, SelectableTable, ExpandableTable } from './table';
 export { OverflowText } from './overflow-text';
 export { ItemsActions } from './items-actions';
 export { Badge } from './badge';
 export { Modal } from './modal';
-export { TableCheckbox } from './table-checkbox';
 export { Divider } from './divider';
-export { Dropdown } from './dropdown';
 export { TabsPanel, Tab } from './tabs';
 export { MessagePanel } from './message-panel';
 export { Spinner } from './spinner';
-export { FormGroup } from './form-group';
-export { Textarea } from './textarea';
-export { DropdownInput } from './dropdown-input';
