--- src/components/table-checkbox/index.ts
@@ -1,1 +1,0 @@
-export { TableCheckbox } from './table-checkbox';
