+++ src/components/table/selectable-table.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 
+import { Inputs } from '../../forms';
 import { Table } from './table';
 import { Column } from './column';
-import { TableCheckbox } from '../table-checkbox';
 
 interface Props {
-  data: object[];
+  data: any[];
   children: React.ReactNode;
   idKey: string;
   footer?: React.ReactNode;
@@ -31,7 +31,7 @@ export const SelectableTable = ({
         name="selector"
         Cell={({ item }) => {
           return (
-            <TableCheckbox
+            <Inputs.Checkbox
               onClick={() => {
                 selectedRows.includes(item[idKey])
                   ? onSelect(selectedRows.filter((selectedItem) => selectedItem !== item[idKey]))
@@ -42,7 +42,7 @@ export const SelectableTable = ({
           );
         }}
         HeaderCell={() => (
-          <TableCheckbox
+          <Inputs.Checkbox
             onClick={() => {
               onSelect(!isAllSelected ? data.map((item: any) => String(item[idKey])) : []);
             }}
