+++ src/forms/index.ts
@@ -1,2 +1,4 @@
+export { Inputs, Button, DropdownInput } from './inputs';
 export { Fields } from './fields';
+export { FormGroup } from './form-group';
 export { composeValidators, required, sizeLimit } from './form-validators';
