+++ src/forms/inputs/button/button.module.scss
@@ -1,4 +1,4 @@
-@import '../../common/style-guide/index';
+@import '../../../common/style-guide/index';
 
 .button {
   border-radius: 20px;
