+++ src/forms/inputs/dropdown-input/dropdown-input.module.scss
@@ -1,4 +1,4 @@
-@import '../../common/style-guide/index';
+@import '../../../common/style-guide/index';
 
 .dropdown-input {
   position: relative;
