+++ src/forms/inputs/dropdown-input/dropdown-item.ts
@@ -1,0 +1,4 @@
+export interface DropdownItem {
+  value: string;
+  label: string;
+}
