+++ src/forms/inputs/dropdown/dropdown.module.scss
@@ -1,4 +1,4 @@
-@import '../../common/style-guide/index';
+@import '../../../common/style-guide/index';
 
 .dropdown {
   position: relative;
