+++ src/forms/inputs/index.ts
@@ -1,0 +1,15 @@
+import { Input } from './input';
+import { Dropdown } from './dropdown';
+import { Textarea } from './textarea';
+import { Checkbox } from './checkbox';
+import { Toggler } from './toggler';
+
+export { DropdownInput } from './dropdown-input';
+export { Button } from './button';
+export const Inputs = {
+  Text: Input,
+  Dropdown,
+  Textarea,
+  Checkbox,
+  Toggler,
+};
