+++ src/forms/inputs/input/input.module.scss
@@ -1,4 +1,4 @@
-@import '../../common/style-guide/index';
+@import '../../../common/style-guide/index';
 
 .input-wrapper {
   position: relative;
