+++ src/forms/inputs/textarea/textarea.module.scss
@@ -1,4 +1,4 @@
-@import '../../common/style-guide/index';
+@import '../../../common/style-guide/index';
 
 .textarea {
   width: 100%;
