+++ src/forms/inputs/toggler/toggler.module.scss
@@ -1,4 +1,4 @@
-@import '../../common/style-guide/index';
+@import '../../../common/style-guide/index';
 
 .toggler {
   position: relative;
